Crate page

This crate provides abstractions and implementations for data and index pages.

Modules

errors 🔒

header

Page Header Definition and Accessors

insertion_plan

Defines the insertion plan struct for inserting records into unsorted heap pages, along with related enums.

page

Slotted structure and related functionality. Module defining a fixed-size slotted page structure with its associated methods.

page_id

Unique identifier for pages.

page_type

Different types of pages supported.

slot 🔒

slot_array 🔒

Slot Array Handling for Slotted Pages

tests 🔒

Constants

HEADER_SIZE 🔒

Size of the header in bytes.

PAGE_SIZE

Fixed-size of a page in bytes