page::header

Struct HeaderRef

pub(crate) struct HeaderRef<'a> {
    bytes: &'a [u8; 96],
}

The ‘HeaderRef’ struct provides an immutable view into the header of a page.

Fields
bytes: &'a [u8; 96]
Implementations

impl<'a> HeaderRef<'a>

pub fn new(bytes: &'a [u8]) -> Result<Self, HeaderError>

Creates a new HeaderRef from a slice of bytes if it matches the required size.

Parameters

bytes: A reference to a slice of bytes that will be used to initialize the HeaderRef.

Returns

Some(HeaderRef) if the length of the bytes slice is equal to HEADER_SIZE.

None if the length of the bytes slice does not match HEADER_SIZE.

pub fn needs_compaction(&self, row_size: usize) -> Result<bool, HeaderError>

Returns a boolean indicating whether the page needs compaction. Mostly a helper method, since it only uses the values from the header to determine this. Returns Some(bool) if successful or HeaderError if something goes wrong.

impl<'a> HeaderRef<'a>

pub(crate) fn get_slot_count(&self) -> Result<u16, HeaderError>

Getter for field slot_count. Type: u16. Offset: 0.

impl<'a> HeaderRef<'a>

pub(crate) fn get_free_start(&self) -> Result<u16, HeaderError>

Getter for field free_start. Type: u16. Offset: 2.

impl<'a> HeaderRef<'a>

pub(crate) fn get_free_end(&self) -> Result<u16, HeaderError>

Getter for field free_end. Type: u16. Offset: 4.

impl<'a> HeaderRef<'a>

pub(crate) fn get_free_space(&self) -> Result<u16, HeaderError>

Getter for field free_space. Type: u16. Offset: 6.

impl<'a> HeaderRef<'a>

pub(crate) fn get_can_compact(&self) -> Result<u16, HeaderError>

Getter for field can_compact. Type: u16. Offset: 8.

impl<'a> HeaderRef<'a>

pub(crate) fn get_page_number(&self) -> Result<u32, HeaderError>

Getter for field page_number. Type: u32. Offset: 10.

impl<'a> HeaderRef<'a>

pub(crate) fn get_page_type(&self) -> Result<u16, HeaderError>

Getter for field page_type. Type: u16. Offset: 14.

impl<'a> HeaderRef<'a>

pub(crate) fn get_left_page(&self) -> Result<u32, HeaderError>

Getter for field left_page. Type: u32. Offset: 16.

impl<'a> HeaderRef<'a>

pub(crate) fn get_right_page(&self) -> Result<u32, HeaderError>

Getter for field right_page. Type: u32. Offset: 20.

impl<'a> HeaderRef<'a>

pub(crate) fn get_last_lsn(&self) -> Result<u64, HeaderError>

Getter for field last_lsn. Type: u64. Offset: 26.