page::impls

Struct Page

pub struct Page {
    page_id: PageId,
    data: Box<[u8; 4096]>,
}

Wrapper around a fixed-size byte array representing a page.

Fields
page_id: PageId

Unique identifier of the page. Comprised of file_name_hash::page_number_within_file

data: Box<[u8; 4096]>

Main binary array holding the PAGE_SIZE bytes of data for the page. Boxed and owned by this struct.

Implementations

impl Page

Accessor methods for the Page struct.

pub fn page_id(&self) -> PageId

Returns the unique identifier of the page.

impl Page

Methods for creating and initializing pages.

pub(crate) fn new_zeroed(page_id: PageId) -> Self

Creates a new page with all bytes initialized to zero. Private constructor.

pub(crate) fn new_from_bytes(bytes: Box<[u8; 4096]>, page_id: PageId) -> Self

Creates a new page from an existing byte array.

pub(crate) fn new_empty( page_id: PageId, page_type: PageType, ) -> Result<Self, HeaderError>

Creates a new empty page with the specified page ID and page type.

impl Page

pub(super) fn delete_row_internal( &mut self, slot_index: usize, compact_requested: bool, ) -> Result<(), DeleteError>

impl Page

Header access methods for the Page struct.

pub(crate) fn header_ref(&self) -> Result<HeaderRef<'_>, HeaderError>

Returns a read-only reference to the page header.

pub(crate) fn header_mut(&mut self) -> Result<HeaderMut<'_>, HeaderError>

Returns a mutable reference to the page header.

impl Page

pub(super) fn insert_row_unsorted_internal( &mut self, plan: InsertionPlan, bytes: Vec<u8>, ) -> Result<(), InsertError>

impl Page

pub(super) fn plan_insert_internal( &self, row_len: usize, ) -> Result<InsertionPlan, InsertError>

fn get_insertion_slot(&self) -> Result<InsertionSlot, InsertError>

fn find_insertion_offset( &self, row_len: usize, ) -> Result<InsertionOffset, InsertError>

impl Page

Private methods for the Page struct.

pub(crate) fn slot_array_ref(&self) -> Result<SlotArrayRef<'_>, SlotError>

Returns an immutable view of the slot array.

pub(crate) fn slot_array_mut(&mut self) -> Result<SlotArrayMut<'_>, SlotError>

Returns a mutable view of the slot array.

pub(super) fn row_size_fits(&self, row_size: usize) -> Result<bool, HeaderError>

Determines whether the requested row size fits on the page. Does not account for fragmentation (i.e., row might fit only after a compaction of the page). Returns a boolean or error if something goes wrong while processing the header.

pub(super) fn is_slot_valid( &self, slot: &SlotRef<'_>, ) -> Result<bool, SlotError>

Determines if a slot is valid (used) or it has been invalidated (row referenced by it was deleted).

pub(super) fn compact(&mut self) -> Result<(), SlotError>

impl Page

Internal row access methods for the Page struct.

pub(super) fn read_row_internal( &self, slot_index: u32, ) -> Result<&[u8], ReadRowError>

Retrieves a row by its slot index. Returns a slice of bytes representing the row data.

impl Page

Public APIs for the Page struct. All public APIs use the PageResult type

pub fn row(&self, slot_index: u32) -> Result<&[u8], PageError>

Retrieves a row from the page by its slot index.

Arguments

slot_index - The index of the slot to retrieve the row from. Indexing starts from 0.

Returns

PageResult<&[u8]> - A result containing a reference to the row data as a byte slice if successful, or an error wrapped in PageResult if the operation fails.

Errors

This method can return the following errors:

PageError - If there is an issue with the operation, such as an invalid slot index.

The error is augmented with the page_id of the current page for better traceability.

pub fn plan_insert(&self, row_len: usize) -> Result<InsertionPlan, PageError>

Plans the insertion of a row into the page. Used only for heap pages.

Arguments

row_len - The length of the row to be inserted, in bytes.

Returns

PageResult<InsertionPlan> - A result containing the insertion plan if successful, or an error wrapped in PageResult if the operation fails.

Errors

This method can return the following errors:

PageError - If there is an issue with the operation, such as insufficient space or other constraints preventing the insertion.

The error is augmented with the page_id of the current page for better traceability.

pub fn insert_heap( &mut self, plan: InsertionPlan, row: Vec<u8>, ) -> Result<(), PageError>

Inserts a row into a heap page using the provided insertion plan.

Arguments

plan - The InsertionPlan that specifies where and how the row should be inserted.

row - A Vec<u8> containing the row data to be inserted.

Returns

PageResult<()> - A result indicating success (Ok(())) or failure (Err(PageError)).

Errors

This method can return the following errors:

PageOpError - If there is an issue during the insertion process. PageOpError will contain the source error.

The error is augmented with the page_id of the current page for better traceability.

pub fn delete_row( &mut self, slot_index: usize, compact_requested: bool, ) -> Result<(), PageError>

Deletes a row from the page at the specified slot index.

Arguments

slot_index - The index of the slot from which the row should be deleted. Indexing starts from 0.

compact_requested - A boolean flag indicating whether compaction should be performed after the deletion. If true, the page may attempt to compact the remaining rows to reclaim space.

Returns

PageResult<()> - A result indicating success (Ok(())) or failure (Err(PageOpError)). The error is augmented with the page_id of the current page for better traceability.

Errors

This method can return the following errors:

PageOpError - If there is an issue during the deletion process, such as an invalid slot index or other constraints preventing the deletion.